+++ rets/__init__.py
@@ -2,8 +2,8 @@ from .session import Session
 from .exceptions import RETSException
 
 __title__ = 'rets'
-__version__ = '0.3.0'
+__version__ = '0.3.2'
 __author__ = 'REfindly'
 __license__ = 'MIT'
-__copyright__ = 'Copyright 2016 REfindly'
+__copyright__ = 'Copyright 2017 REfindly'
 __all__ = ['Session', 'RETSException']